renaming filewriter_command_dialog.py to filewriter_command_widget.py
